aede_core::tags

Module id3

ID3v2 (versions 2.2, 2.3 and 2.4) and ID3v1.

This module serves MP3, but also WAV, AIFF and some FLAC files, all of which may embed an ID3v2 block.

Reference: https://id3.org/id3v2.4.0-structure

Functions

read_id3v1

Reads the ID3v1 tag from the last 128 bytes. Used only as a fallback, when ID3v2 is absent or silent about a field.

read_id3v2

Reads the ID3v2 tag at the head of the file, if any.

skip_id3v2

Returns the offset of the first byte located after a possible ID3v2 tag, that is the real start of the audio stream. Returns 0 when there is no tag.